feat(chat): refactor MessageList component to improve readability and maintainability by extracting MessageItem and DeleteMessageDialog components fix(chat): enhance streaming message handling with significant content change detection for smoother user experience chore(chat): remove unused imports and code to clean up the MessageList component feat(chat): implement memoization for performance optimization in MessageItem and MessageContent components fix(hooks): update streaming content using requestAnimationFrame for smoother UI updates in useSendMessage hook
